app/pyproject.toml
[tool.poetry]
name = "plugger"
version = "0.1.0"
description = ""
authors = ["Julen Badiola <badiolajulen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.11"
gunicorn = "20.1.0"
Unipath = "1.1"
whitenoise = "5.3.0"
python-decouple = "3.4"
docker = "5.0.3"
PyYAML = "^6.0"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^6.4.1"
selenium = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
